Another attempt to cleanup the solution directory retrieval
